Contract multiple \e[m "breakers"
It is totally legal for the `reset` ANSI sequence to be specified
multiple times in a row. This does not mean that we should add more than
one `</tspan>`, though.

Note: this is _still_ not quite correct. For example, if `bold` mode is
turned on, then some text is written, then the color is switched to
`cyan`, another couple characters are written, a subsequent `reset`
would need to close _two_ `<tspan>`s.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>